Small formatting changes to make pull request more readable with less changes Yes this adds some whitespace that is currently in the base but it makes it easier to read the changes because it doesn't show the dif for removing the spacing

+++ src/test/java/moze_intel/projecte/emc/mappers/customConversions/CustomConversionMapperTest.java
@@ -111,9 +111,9 @@ public class CustomConversionMapperTest
 		CustomConversionFile f = CustomConversionMapper.parseJson(new StringReader(simpleFile));
 		assertNotNull(f.values);
 		assertEquals(1, f.values.setValueBefore.get(NSSItem.create("a", 0)).longValue());
-		assertEquals(2, (int) f.values.setValueBefore.get(NSSItem.create("b", 0)).longValue());
+		assertEquals(2, f.values.setValueBefore.get(NSSItem.create("b", 0)).longValue());
 		assertEquals(Long.MIN_VALUE, f.values.setValueBefore.get(NSSItem.create("c", 0)).longValue());
-		assertEquals(3, (int) f.values.setValueAfter.get(NSSItem.create("d", 0)).longValue());
+		assertEquals(3, f.values.setValueAfter.get(NSSItem.create("d", 0)).longValue());
 
 	}
